is public

+++ controllers/api.js
@@ -1,10 +1,17 @@
 const Sequelize = require('sequelize');
-const isProduction = process.env.NODE_ENV === 'production';
-const config = isProduction ? require('../config_online') : require('../config');
-
+var moment = require('moment');
 const passport = require('../utils/auth');
 
-console.log('init sequelize...');
+const ERROR_MSG_NOT_LOGIN = 'not login';
+const ERROR_MSG_NO_CEPT_AUTH = 'no cept auth';
+
+const CEPT_OP_TYPE_READ = 1;
+const CEPT_OP_TYPE_UPDATE = 2;
+
+const isProduction = process.env.NODE_ENV === 'production';
+const config = isProduction ? require('../config_online') : require('../config');
+const DATE_FORMAT = 'YYYY-MM-DD HH:mm:ss';
+console.log(moment().local().format(DATE_FORMAT));
 
 var sequelize = new Sequelize(config.database, config.username, config.password, {
     host: config.host,
@@ -23,12 +30,15 @@ var ConceptModel = sequelize.define('concept', {
         autoIncrement: true   // in order to get the new created record's id after ConceptModel.create done
     },
     name: Sequelize.STRING(64),
-    description: Sequelize.TEXT
+    description: Sequelize.TEXT,
+    user_id: Sequelize.BIGINT,
+    is_public: Sequelize.BIGINT(2),
+    created: Sequelize.DATE
 }, {
-        freezeTableName: true, // 默认false修改表名为复数，true不修改表名，与数据库表名同步
-        tableName: 'concept',
-        timestamps: false
-    });
+    freezeTableName: true, // 默认false修改表名为复数，true不修改表名，与数据库表名同步
+    tableName: 'concept',
+    timestamps: false
+});
 
 var ConceptContextModel = sequelize.define('concept_context', {
     id: {
@@ -39,12 +49,36 @@ var ConceptContextModel = sequelize.define('concept_context', {
     cid: Sequelize.BIGINT,
     context_concept_id: Sequelize.BIGINT
 }, {
-        freezeTableName: true, // 默认false修改表名为复数，true不修改表名，与数据库表名同步
-        tableName: 'concept_context',
-        timestamps: false
-    });
+    freezeTableName: true, // 默认false修改表名为复数，true不修改表名，与数据库表名同步
+    tableName: 'concept_context',
+    timestamps: false
+});
 
-async function getConceptContext(cid, cidDist) {
+async function verifyCeptOpAuth(cid, uid, opType = 0) {
+    var cept = await getConceptDetail(cid);
+    if (cept) {
+        switch(opType) {
+            case CEPT_OP_TYPE_READ:
+                if (cept.user_id === uid || cept.is_public) {
+                    return true;
+                }
+                break;
+            case CEPT_OP_TYPE_UPDATE:
+                if (cept.user_id === uid) {
+                    return true;
+                }
+                break;
+            default:
+                return false;
+                break;
+        }
+    }
+
+    return false;
+}
+
+// cidDist && uid 都是注入的辅助变量
+async function getConceptContext(cid, cidDist, uid = 0) {
     let ret = [];
 
     var conceptContexts = await ConceptContextModel.findAll({
@@ -56,10 +90,19 @@ async function getConceptContext(cid, cidDist) {
     if (conceptContexts && conceptContexts.length > 0) {
         for (let i = 0; i < conceptContexts.length; i++) {
             let item = conceptContexts[i];
+
+            var verifyResult = await verifyCeptOpAuth(item.context_concept_id, uid, CEPT_OP_TYPE_READ);
+            if (!verifyResult) {
+                item.dirty = true;
+                cidDist[item.context_concept_id] = -1;  // dirty
+                ret.push(item);
+                continue;
+            }
+
             ret.push(item);
             if (!cidDist[item.context_concept_id]) {
                 cidDist[item.context_concept_id] = 1;
-                let subRet = await getConceptContext(item.context_concept_id, cidDist);
+                let subRet = await getConceptContext(item.context_concept_id, cidDist, uid);
                 ret = ret.concat(subRet);
             }
         }
@@ -79,26 +122,33 @@ async function getConceptDetail(cid) {
     return concept;
 }
 
-async function getConceptDetailByQuery(query) {
-    let ret = [];
-
-    var concept = await ConceptModel.findOne({
-        where: {
-            name: query
-        }
-    });
-
-    return concept;
-}
-
 module.exports = {
     'GET /api/getAllCept': async (ctx, next) => {
-        var ret = await ConceptModel.findAll();
+        var user = ctx.state.user;
+        var userId = user ? user.id : 0;
+
+        var ret = await ConceptModel.findAll({
+            'where': {
+                '$or': [
+                    {'user_id': userId},
+                    {'is_public': 1}
+                ]
+            }
+        });
         ctx.rest(ret);
     },
 
     'GET /api/getCeptById/:id': async (ctx, next) => {
+        var user = ctx.state.user;
+        var userId = user ? user.id : 0;
         var cid = parseInt(ctx.params.id);
+
+        var verifyResult = await verifyCeptOpAuth(cid, userId, CEPT_OP_TYPE_READ);
+        if (!verifyResult) {
+            ctx.rest({status: 0, msg: ERROR_MSG_NO_CEPT_AUTH});
+            return;
+        }
+
         var ret = {
             detail: null,
             ctx: null,
@@ -109,18 +159,36 @@ module.exports = {
             ret.detail = ceptDetail;
             var cidDist = {};
             cidDist[cid] = 1;
-            ret.ctx = await getConceptContext(cid, cidDist);
+            ret.ctx = await getConceptContext(cid, cidDist, userId);
             for (var key in cidDist) {
                 var cd = await getConceptDetail(key);
-                ret.ctxInfoList[key] = cd.name;
+                var item = {
+                    id: cd.id,
+                    name: cd.name
+                }
+                if (cidDist[cd.id] === -1) {
+                    item.dirty = true;
+                }
+                ret.ctxInfoList[key] = item;
             }
         }
         ctx.rest(ret);
     },
 
     'GET /api/getCeptByQuery/query/:query': async (ctx, next) => {
+        var user = ctx.state.user;
+        var userId = user ? user.id : 0;
+
         var query = ctx.params.query;
-        var ret = await getConceptDetailByQuery(query);
+        var ret = await ConceptModel.findOne({
+            where: {
+                'name': query,
+                '$or': [
+                    {'user_id': userId},
+                    {'is_public': 1}
+                ]
+            }
+        });
         ctx.rest(ret);
     },
 
@@ -131,33 +199,72 @@ module.exports = {
     },
 
     'POST /api/createCept': async (ctx, next) => {
+        if (!ctx.isAuthenticated()) {
+            ctx.rest({status: 0, msg: ERROR_MSG_NOT_LOGIN});
+            return;
+        }
+
+        var user = ctx.state.user;
         var name = ctx.request.body.name;
         var description = ctx.request.body.description;
+        var isPublic = parseInt(ctx.request.body.is_public);
+        var currentDate = moment().local().format(DATE_FORMAT);
         var ret = await ConceptModel.create({
             name: name,
-            description: description
+            description: description,
+            user_id: parseInt(user.id),
+            is_public: isPublic,
+            created: currentDate
         });
 
         ctx.rest(ret);
     },
 
     'POST /api/updateCept': async (ctx, next) => {
+        if (!ctx.isAuthenticated()) {
+            ctx.rest({status: 0, msg: ERROR_MSG_NOT_LOGIN});
+            return;
+        }
+
+        var user = ctx.state.user;
         var cid = parseInt(ctx.request.body.cid);
+
+        var verifyResult = await verifyCeptOpAuth(cid, user.id, CEPT_OP_TYPE_UPDATE);
+        if (!verifyResult) {
+            ctx.rest({status: 0, msg: ERROR_MSG_NO_CEPT_AUTH});
+            return;
+        }
+
         var description = ctx.request.body.description;
+        var isPublic = parseInt(ctx.request.body.is_public);
         var ret = await ConceptModel.update({
-            description: description
+            description: description,
+            is_public: isPublic
         }, {
             where: {
                 id: cid
             }
         });
 
-        ctx.rest(ret);
+        ctx.rest({status: 1, data: ret});
     },
 
     'POST /api/addCtx': async (ctx, next) => {
+        if (!ctx.isAuthenticated()) {
+            ctx.rest({status: 0, msg: ERROR_MSG_NOT_LOGIN});
+            return;
+        }
+
         var ret = {};
+        var user = ctx.state.user;
         var cid = parseInt(ctx.request.body.cid);
+
+        var verifyResult = await verifyCeptOpAuth(cid, user.id, CEPT_OP_TYPE_UPDATE);
+        if (!verifyResult) {
+            ctx.rest({status: 0, msg: ERROR_MSG_NO_CEPT_AUTH});
+            return;
+        }
+
         var name = ctx.request.body.name;
         var cept = await ConceptModel.findOne({
             where: {
@@ -171,23 +278,44 @@ module.exports = {
         });
 
         if (cept && ceptCtx) {
+            var verifyResult = await verifyCeptOpAuth(ceptCtx.id, user.id, CEPT_OP_TYPE_READ);
+            if (!verifyResult) {
+                ctx.rest({status: 0, msg: ERROR_MSG_NO_CEPT_AUTH});
+                return;
+            }
+
             ret = await ConceptContextModel.create({
                 cid: cid,
                 context_concept_id: ceptCtx.id
             });
         }
-        ctx.rest(ret);
+        ctx.rest({status: 1, data: ret});
     },
 
     'POST /api/removeCtx': async (ctx, next) => {
+        if (!ctx.isAuthenticated()) {
+            ctx.rest({status: 0, msg: ERROR_MSG_NOT_LOGIN});
+            return;
+        }
+
+        var user = ctx.state.user;
+        var cid = parseInt(ctx.request.body.cid);
+
+        var verifyResult = await verifyCeptOpAuth(cid, user.id, CEPT_OP_TYPE_UPDATE);
+        if (!verifyResult) {
+            ctx.rest({status: 0, msg: ERROR_MSG_NO_CEPT_AUTH});
+            return;
+        }
+
         var ret = ConceptContextModel.destroy({
             where: {
-                cid: parseInt(ctx.request.body.cid),
+                cid: cid,
                 context_concept_id: parseInt(ctx.request.body.ccid)
             }
         });
         ctx.rest(ret);
     },
+
 
     // login
     'POST /api/login': async (ctx, next) => {
@@ -208,5 +336,6 @@ module.exports = {
 
     'GET /api/userInfo': async (ctx, next) => {
         ctx.rest(ctx.state.user);
+        // ctx.rest({status: ctx.isAuthenticated()});
     }
 };
